PRO-3267 WIP commit to get applying not applying executor list changes persisted
