+++ src/main/java/uk/gov/hmcts/probate/model/ccd/raw/request/CaseData.java
@@ -17,7 +17,6 @@ import uk.gov.hmcts.probate.model.ccd.raw.AliasName;
 import uk.gov.hmcts.probate.model.ccd.raw.CollectionMember;
 import uk.gov.hmcts.probate.model.ccd.raw.Document;
 import uk.gov.hmcts.probate.model.ccd.raw.DocumentLink;
-import uk.gov.hmcts.probate.model.ccd.raw.ProbateAliasName;
 import uk.gov.hmcts.probate.model.ccd.raw.SolsAddress;
 import uk.gov.hmcts.probate.model.ccd.raw.StopReason;
 
@@ -89,7 +88,7 @@ public class CaseData {
 
     private final List<CollectionMember<AliasName>> solsDeceasedAliasNamesList;
 
-    private final List<CollectionMember<ProbateAliasName>> BODeceasedAliasNamesList;
+   // private final List<CollectionMember<ProbateAliasName>> BODeceasedAliasNamesList;
 
     @NotBlank(groups = {ApplicationUpdatedGroup.class}, message = "{solsIHTFormIdIsNull}")
     private final String solsIHTFormId;
@@ -192,9 +191,9 @@ public class CaseData {
     @DecimalMin(groups = {NextStepsConfirmationGroup.class}, value = "0.0", message = "{totalFeeNegative}")
     private final BigDecimal totalFee;
 
-    private List<CollectionMember<AdditionalExecutorApplying>> executorsApplying;
+    private List<CollectionMember<AdditionalExecutorApplying>> additionalExecutorsApplying;
 
-    private List<CollectionMember<AdditionalExecutorNotApplying>> executorsNotApplying;
+    private List<CollectionMember<AdditionalExecutorNotApplying>> additionalExecutorsNotApplying;
 
     private final ApplicationType applicationType;
 
@@ -213,13 +212,13 @@ public class CaseData {
 
     private final String boLimitationText;
 
-//    public List<CollectionMember<AdditionalExecutor>> getExecutorsApplying() {
-//        return getAllExecutors(true);
-//    }
-//
-//    public List<CollectionMember<AdditionalExecutor>> getExecutorsNotApplying() {
-//        return getAllExecutors(false);
-//    }
+    public List<CollectionMember<AdditionalExecutor>> getExecutorsApplying() {
+        return getAllExecutors(true);
+    }
+
+    public List<CollectionMember<AdditionalExecutor>> getExecutorsNotApplying() {
+        return getAllExecutors(false);
+    }
 
     public boolean isPrimaryApplicantApplying() {
         return YES.equals(primaryApplicantIsApplying);
