+++ src/main/java/uk/gov/hmcts/probate/model/ccd/raw/response/ResponseCaseData.java
@@ -11,7 +11,6 @@ import uk.gov.hmcts.probate.model.ccd.raw.AliasName;
 import uk.gov.hmcts.probate.model.ccd.raw.CollectionMember;
 import uk.gov.hmcts.probate.model.ccd.raw.Document;
 import uk.gov.hmcts.probate.model.ccd.raw.DocumentLink;
-import uk.gov.hmcts.probate.model.ccd.raw.ProbateAliasName;
 import uk.gov.hmcts.probate.model.ccd.raw.SolsAddress;
 import uk.gov.hmcts.probate.model.ccd.raw.StopReason;
 
@@ -79,7 +78,7 @@ public class ResponseCaseData {
     private final String deceasedAnyOtherNames;
     private final SolsAddress primaryApplicantAddress;
     private final List<CollectionMember<AliasName>> solsDeceasedAliasNamesList;
-    private final List<CollectionMember<ProbateAliasName>> BODeceasedAliasNamesList;
+    //private final List<CollectionMember<ProbateAliasName>> BODeceasedAliasNamesList;
     private final String ccdState;
     private final String casePrinted;
     private final String boEmailDocsReceivedNotificationRequested;
