+++ src/test/java/uk/gov/hmcts/probate/transformer/CallbackResponseTransformerTest.java
@@ -175,8 +175,8 @@ public class CallbackResponseTransformerTest {
                 .casePrinted(CASE_PRINT)
                 .boCaseStopReasonList(STOP_REASONS_LIST)
                 .willExists(YES)
-                .executorsApplying(ADDITIONAL_EXEC_LIST_APP)
-                .executorsNotApplying(ADDITIONAL_EXEC_LIST_NOT_APP)
+                .additionalExecutorsApplying(ADDITIONAL_EXEC_LIST_APP)
+                .additionalExecutorsNotApplying(ADDITIONAL_EXEC_LIST_NOT_APP)
                 .boDeceasedTitle(DECEASED_TITLE)
                 .boDeceasedHonours(DECEASED_HONOURS)
                 .boWillMessage(WILL_MESSAGE)
